requirements/dev.txt
-r base.txt

# --- Testing ---
pytest==8.2.1
pytest-django==4.8.0
pytest-cov==5.0.0
factory-boy==3.3.0
faker==25.2.0

# --- Mocks ---
moto[s3]==5.0.7

# --- Tooling ---
django-debug-toolbar==4.3.0
ruff==0.4.4
ipython==8.24.0

# --- Celery monitoring (UI) ---
flower==2.0.1
